Adds select_runnable decorator to api and bumps to version 0.3.4
The select_runnable decorator allows the user to pick which runnable is
run for each type of file validator. This means that you can run
JavaScript unit tests when '.js' files are modified and Python tests
when you edit a '.py' file. When the select_runnable decorator is used
the behaviour of sniffer changes slightly in that all file validators
are attempted and if one of them returns True the associated runnable is
then executed.